+++ hassio-addon/CHANGELOG.md
@@ -2,6 +2,11 @@
 - 한달정도 특별한 오류가 없으면 beta 딱지를 떼고 stable로 승격시킬 예정입니다.
 - 오류가 있으면 homeasssutant 카페, 깃헙, 디스코드 등으로 알려주세요.
 
+v2.8.3
+- fix: 개별 docker환경에서 CONFIG_FILES를 비워두면(권장) 앱이 시작되지 않는 현상 수정
+- fix: 특정상황에서 portId를 잘못 찾는 문제 수정
+  - `[service] findConfigIndexForEntity called for ${entityId}. This is an exceptional situation where portId was not found.` 이런식의 에러가 로그에 있다면 알려주세요.
+
 v2.8.1-2.8.2
 - arm64 빌드 실패 수정